server/requirements.txt
Flask==2.3.3
pymongo==4.7.0
flask-cors==4.0.0
python-dotenv==1.0.0
werkzeug==2.3.7
